+++ src/classic/hfp_ag.c
@@ -910,7 +910,7 @@ static void hfp_ag_emit_general_simple_event(uint8_t event_subtype){
 
 static void hfp_ag_emit_apple_accessory_information(hfp_connection_t *hfp_connection){
     uint8_t event[40];
-    uint16_t version_len = strlen(hfp_connection->apple_accessory_version);
+    uint16_t version_len = (uint16_t)strlen(hfp_connection->apple_accessory_version);
     event[0] = HCI_EVENT_HFP_META;
     event[1] = 9 + version_len;
     event[2] = HFP_SUBEVENT_APPLE_ACCESSORY_INFORMATION;
@@ -918,7 +918,7 @@ static void hfp_ag_emit_apple_accessory_information(hfp_connection_t *hfp_connec
     little_endian_store_16(event, 5, hfp_connection->apple_accessory_vendor_id);
     little_endian_store_16(event, 7, hfp_connection->apple_accessory_product_id);
     event[9] = hfp_connection->apple_accessory_features;
-    event[10] = version_len;
+    event[10] = (uint8_t)version_len;
     memcpy(&event[11], hfp_connection->apple_accessory_version, version_len);
     (*hfp_ag_callback)(HCI_EVENT_PACKET, 0, event, 11 + version_len);
 }
